feat: added new steps
