+++ admin/src/onboarding/Header.jsx
@@ -15,7 +15,8 @@ const Header = () => {
 			plan_choice: __( 'Get Started' ),
 			api_key: __( 'Add your API key' ),
 			schedule: __( 'Add your domain' ),
-			seo_setup: __( 'Choose your competitors' ),
+			choose_keywords: __( 'Choose keywords' ),
+			choose_competitors: __( 'Choose competitors' ),
 			modules: __( 'Activate modules' ),
 		};
 	}, [ __ ] );
