+++ admin/src/onboarding/steps/StepChooseCompetitors.jsx
